Initial add of FormatterFactory
